removed old code and refined mpu6050

+++ Lokaverkefni/transmitter/Coordinate.h
@@ -17,7 +17,7 @@ class Coordinate
         void correct_orientation();
         void update_data(int new_x, int new_y, bool new_button_state=1);
         void debug();
-        uint8_t* get_positional_data(uint8_t *arr, uint8_t to_max=255, int a_max=255, int f_max=1, int b_max=1, int r_max=1, int l_max=1);
+        uint8_t* get_positional_data(uint8_t *arr, float to_max=255, int a_max=255, int f_max=1, int b_max=1, int r_max=1, int l_max=1);
         uint8_t _orientation;
     private:
         int _x;
